// src/enums/actions.ts
export enum Actions {
    CLOSE_COMPLETE_GAME_POPUP = 'CLOSE_COMPLETE_GAME_POPUP',
    EXPAND_MENU_PANEL = 'EXPAND_MENU_PANEL',
    GAME_BOARD_SIZE_CHANGED = 'GAME_BOARD_SIZE_CHANGED',
    GET_CARD = 'GET_CARD',
    NEXT_CARD = 'NEXT_CARD',
    OPEN_COMPLETE_GAME_POPUP = 'OPEN_COMPLETE_GAME_POPUP',
    PUT_CARD_TO_DECK = 'PUT_CARD_TO_DECK',
    RESTART_GAME = 'RESTART_GAME',
    SELECT_CARD = 'SELECT_CARD',
    SELECT_LANG = 'SELECT_LANG',
    SWITCH_RESTART_CONFIRMATION_POPUP = 'SWITCH_RESTARTE_CONFIRMATION_POPUP',
    UPDATE_CARDS_STATE = 'UPDATE_CARDS_STATE',
}